refactor(registrar): move orphan deregistration tests
